test: enforce flat-test conventions across the repo
Splits five sibling-describe files into one-describe-per-file:
- core/test/load.test.ts → + load-validation.test.ts
- core/test/load-layered.test.ts → + load-layered-validation.test.ts
- core/test/presets.test.ts → + load-presets.test.ts
- addon/test/virtual-plugin.test.ts → + integration-side-effects.test.ts
- blocks/test/color-table.test.tsx → + color-table-grouping.test.tsx
                                    + color-table-expansion.test.tsx
  (color-table tests share `_color-table-helpers.tsx`)

Annotates the seven previously-unannotated `beforeAll` blocks
(core variance / chrome / emit-via-terrazzo / load / load-presets;
integrations tailwind / css-in-js) with the perf-escape-hatch
reason: each shares a ~1s loadProject across every assertion in
the file.

Annotates the resolver-edge-cases `beforeEach` (per-test tmpdir
mint) so future readers understand the per-test reset is
functional, not cosmetic.

Closes #705
