+++ packages/core/test/load-layered.test.ts
@@ -108,23 +108,3 @@ describe('loadProject — layered axes', () => {
   });
 });
 
-describe('loadProject — config validation', () => {
-  it('throws when both resolver and axes are set', async () => {
-    const config = {
-      tokens: ['base/*.json'],
-      resolver: 'resolver.json',
-      axes: [{ name: 'mode', contexts: { Light: [] }, default: 'Light' }],
-    } satisfies Config;
-    await expect(loadProject(config, fixtureCwd)).rejects.toThrow(
-      /either `resolver` or `axes`, not both/,
-    );
-  });
-
-  it('accepts a tokens-only config and synthesizes a single theme axis', async () => {
-    const project = await loadProject({ tokens: ['base/*.json'] }, fixtureCwd);
-    expect(project.axes).toEqual([
-      { name: 'theme', contexts: ['default'], default: 'default', source: 'synthetic' },
-    ]);
-    expect(project.themes.map((t) => t.name)).toEqual(['default']);
-  });
-});
