+++ packages/core/test/presets.test.ts
@@ -1,11 +1,6 @@
-import { dirname } from 'node:path';
-import { beforeAll, describe, expect, it } from 'vitest';
-import { resolverPath, tokensDir } from '@unpunnyfuns/swatchbook-tokens';
-import { loadProject } from '#/load';
+import { describe, expect, it } from 'vitest';
 import { validatePresets } from '#/presets';
-import type { Axis, Preset, Project } from '#/types';
-
-const fixtureCwd = dirname(tokensDir);
+import type { Axis, Preset } from '#/types';
 
 const axes: Axis[] = [
   {
@@ -82,36 +77,3 @@ describe('validatePresets', () => {
   });
 });
 
-describe('loadProject — presets', () => {
-  let project: Project;
-
-  beforeAll(async () => {
-    project = await loadProject(
-      {
-        tokens: ['tokens/**/*.json'],
-        resolver: resolverPath,
-        default: { mode: 'Light', brand: 'Default', contrast: 'Normal' },
-        presets: [
-          { name: 'Brand A Dark', axes: { mode: 'Dark', brand: 'Brand A' } },
-          { name: 'Default Light', axes: { mode: 'Light' } },
-        ],
-      },
-      fixtureCwd,
-    );
-  }, 30_000);
-
-  it('exposes validated presets on the project', () => {
-    expect(project.presets).toEqual([
-      { name: 'Brand A Dark', axes: { mode: 'Dark', brand: 'Brand A' } },
-      { name: 'Default Light', axes: { mode: 'Light' } },
-    ]);
-  });
-
-  it('defaults to an empty presets array when config.presets is absent', async () => {
-    const p = await loadProject(
-      { tokens: ['tokens/**/*.json'], resolver: resolverPath, default: { mode: 'Light', brand: 'Default' } },
-      fixtureCwd,
-    );
-    expect(p.presets).toEqual([]);
-  });
-});
